+++ src/ios/FirebasePlugin.h
@@ -19,6 +19,7 @@
 - (void)authenticateUserWithGoogle:(CDVInvokedUrlCommand*)command;
 - (void)authenticateUserWithApple:(CDVInvokedUrlCommand*)command;
 - (void)authenticateUserWithMicrosoft:(CDVInvokedUrlCommand*)command;
+- (void)authenticateUserWithFacebook:(CDVInvokedUrlCommand*)command;
 - (void)signInWithCredential:(CDVInvokedUrlCommand*)command;
 - (void)linkUserWithCredential:(CDVInvokedUrlCommand*)command;
 - (void)reauthenticateWithCredential:(CDVInvokedUrlCommand*)command;
@@ -28,12 +29,17 @@
 - (void)reloadCurrentUser:(CDVInvokedUrlCommand*)command;
 - (void)updateUserProfile:(CDVInvokedUrlCommand*)command;
 - (void)updateUserEmail:(CDVInvokedUrlCommand*)command;
+- (void)verifyBeforeUpdateEmail:(CDVInvokedUrlCommand*)command;
 - (void)sendUserEmailVerification:(CDVInvokedUrlCommand*)command;
 - (void)updateUserPassword:(CDVInvokedUrlCommand*)command;
 - (void)sendUserPasswordResetEmail:(CDVInvokedUrlCommand*)command;
 - (void)deleteUser:(CDVInvokedUrlCommand*)command;
 - (void)useAuthEmulator:(CDVInvokedUrlCommand*)command;
 - (void)getClaims:(CDVInvokedUrlCommand*)command;
+- (void)enrollSecondAuthFactor:(CDVInvokedUrlCommand*)command;
+- (void)verifySecondAuthFactor:(CDVInvokedUrlCommand*)command;
+- (void)listEnrolledSecondAuthFactors:(CDVInvokedUrlCommand*)command;
+- (void)unenrollSecondAuthFactor:(CDVInvokedUrlCommand*)command;
 
 // Remote notifications
 - (void)getId:(CDVInvokedUrlCommand*)command;