+++ types/index.d.ts
@@ -158,11 +158,55 @@ export interface FirebasePlugin {
         error?: (err: string) => void
     ): void
     verifyPhoneNumber(
-        success: (value: object) => void,
+        success: (value: object | boolean) => void,
         error: (err: string) => void,
         phoneNumber: string,
-        timeOutDuration: number,
-        fakeVerificationCode?: string
+        opts?: {
+            timeOutDuration: number,
+            fakeVerificationCode: string,
+            requireSmsValidation: boolean
+        },
+    ): void
+    enrollSecondAuthFactor(
+        success: (value: object | boolean) => void,
+        error: (err: string) => void,
+        phoneNumber: string,
+        opts?: {
+            displayName: string,
+            credential: {
+                verificationId: string,
+                code: string
+            },
+            timeOutDuration: number,
+            fakeVerificationCode: string,
+            requireSmsValidation: boolean
+        },
+    ): void
+    verifySecondAuthFactor(
+        success: (value: object | boolean) => void,
+        error: (err: string) => void,
+        params: {
+            selectedIndex?: number,
+            credential?: {
+                verificationId: string,
+                code: string
+            },
+        },
+        opts?: {
+            timeOutDuration: number,
+            fakeVerificationCode: string,
+            phoneNumber: string,
+            requireSmsValidation: boolean
+        },
+    ): void
+    unenrollSecondAuthFactor(
+        success: () => void,
+        error: (err: string) => void,
+        selectedIndex: number
+    ): void
+    listEnrolledSecondAuthFactors(
+        success: (secondFactors: [object]) => void,
+        error: (err: string) => void
     ): void
     setLanguageCode(
         lang: string,
@@ -211,6 +255,11 @@ export interface FirebasePlugin {
         error?: (err: string) => void,
         locale?: string,
     ): void
+    authenticateUserWithFacebook(
+        accessToken: string,
+        success?: (credential:object) => void,
+        error?: (err: string) => void,
+    ): void
     signInWithCredential(
         credential: object,
         success?: () => void,
@@ -255,6 +304,11 @@ export interface FirebasePlugin {
         success?: () => void,
         error?: (err: string) => void
     ): void
+    verifyBeforeUpdateEmail(
+        email: string,
+        success?: () => void,
+        error?: (err: string) => void
+    ): void
     sendUserEmailVerification(
         actionCodeSettings?: {
             handleCodeInApp?: boolean,
@@ -289,6 +343,10 @@ export interface FirebasePlugin {
         host: string,
         port: number,
         success?: () => void,
+        error?: (err: string) => void
+    ): void
+    getClaims(
+        success: (claims: object) => void,
         error?: (err: string) => void
     ): void
     fetch(